refactor: move weather widget to separate component
